requirements.txt
google-cloud-storage==2.18.2
google-cloud-batch==0.17.28
google-cloud-compute==1.19.2
